+++ test/playground/pages/navigate-early/e2e-test.ts
