+++ test/playground/testRun.ts
@@ -3,16 +3,16 @@ export { testRun }
 import path from 'path'
 import { fileURLToPath } from 'url'
 import { testRun as testRunClassic } from '../../examples/react-minimal/.testRun'
-import { testCumulativeSetting } from './pages/about/about.e2e-test'
-import { testSettingEffect } from './pages/config-meta/effect/effects.e2e-test'
-import { testSettingInheritedByDescendants } from './pages/config-meta/cumulative/cumulative.e2e-test'
-import { testSettingOnlyAvailableInCorrectEnv } from './pages/config-meta/env/env.e2e-test'
-import { testDynamicImportFileEnv } from './pages/dynamic-import-file-env/dynamicImportFileEnv.e2e-test'
-import { testSideExports, testRouteStringDefinedInConfigFile } from './pages/markdown-page/sideExports.e2e-test'
-import { testNavigateEarly } from './pages/navigate-early/navigateEarly.e2e-test'
-import { testNestedLayout } from './pages/nested-layout/nestedLayout.e2e-test'
+import { testCumulativeSetting } from './pages/about/e2e-test'
+import { testSettingEffect } from './pages/config-meta/effect/e2e-test'
+import { testSettingInheritedByDescendants } from './pages/config-meta/cumulative/e2e-test'
+import { testSettingOnlyAvailableInCorrectEnv } from './pages/config-meta/env/e2e-test'
+import { testDynamicImportFileEnv } from './pages/dynamic-import-file-env/e2e-test'
+import { testSideExports, testRouteStringDefinedInConfigFile } from './pages/markdown-page/e2e-test'
+import { testNavigateEarly } from './pages/navigate-early/e2e-test'
+import { testNestedLayout } from './pages/nested-layout/e2e-test'
 import { testPrerenderSettings } from './pages/prerender.e2e-test'
-import { testHistoryPushState } from './pages/pushState/history.e2e-test'
+import { testHistoryPushState } from './pages/pushState/e2e-test'
 import { testRedirectMailto } from './pages/redirects.e2e-tests'
 
 const rootDir = path.dirname(fileURLToPath(import.meta.url))
